requirements.txt
aiogram==3.3.0
python-docx==1.1.2
python-dotenv==1.0.1
